+++ lib/index.js
@@ -70,8 +70,12 @@ function Api(options) {
 		}
 	});
 
-	api.express.use(require('./init-passport.js')(api, passport, AnonymousStrategy, JwtStrategy));
-	api.express.use(require('./init-anonymous-user.js')(api));
+	// MIDDLEWARES
+	var initPassport = require('./middlewares/init-passport.js');
+	var initAnonymousUser = require('./middlewares/init-anonymous-user.js');
+
+	api.express.use(initPassport(api, passport, AnonymousStrategy, JwtStrategy));
+	api.express.use(initAnonymousUser(api));
 	api.express.use(bodyParser.json(options['body-parser']));
 
 	// DEFAULT LISTENERS
