+++ lib/middlewares/init-anonymous-user.js
